requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aio-pika==9.5.4
redis[hiredis]==5.2.1
google-generativeai==0.8.4
httpx==0.28.1
pydantic-settings==2.7.1
gspread==6.1.4
google-auth==2.38.0
pyyaml==6.0.2
jinja2==3.1.5
